Added Interactable Raycast & Modelling Overhaul
Massive Update:
Added a raycast Action (delegate) to trigger OnInteract functions in objects. This is a very helpful tool for quickly adding interactivity to assets. Feel free to subscribe to it.
Completely redid Door and some other assets, in terms of repacking in Blender.
Note that I am in the process of decoupling textures and materials from the Wasteland Cabin pack, so you will see duplicated textures in my Interactable Assets folder, and the original imported packaged.

Side notes: made the player collider a bit smaller to fit in doorways.