Practically Complete Ambient AI
AI have working search functions, wander functions, panick behaviors, speeds, persistence (or memory, if you prefer), pathfinder, line of sight, healing, so on and so forth. This is a very complete version that just needs to be configured to attack and defend against the player. Likely that that will be the finalized version as we will be having multiple different enemies.