Pushing my newest AI build to the branch (#7)
* Pushing my newest AI build to the branch

Made a few tweaks to my AI and testing, hoping to push this build to main.

* Hyper-Improved but Buggy AI

Various changes. Better pathfinder and delays coded in, more customization, basic reactions to taking intense damage with a chance to panic, enemies will search for the player where they last saw him. Additionally added basic although very pretty HD Pipeline Renderer and have begun work on enemy projectiles. Experimental phase. May not implement HDPR based on groups opinions on it.

* Practically Complete Ambient AI

AI have working search functions, wander functions, panick behaviors, speeds, persistence (or memory, if you prefer), pathfinder, line of sight, healing, so on and so forth. This is a very complete version that just needs to be configured to attack and defend against the player. Likely that that will be the finalized version as we will be having multiple different enemies.